requirements.txt
streamlit
openai>=1.3.7
faiss-cpu
pypdf
numpy

